examples/citecheck-tutorial/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "citecheck-tutorial"
version = "0.1.0"
description = "Reference implementation of the citecheck CLI from the ACLI tutorial."
requires-python = ">=3.11"
license = "EUPL-1.2"
authors = [{ name = "Alpibru" }]
dependencies = [
    "acli-spec>=0.4.1",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "typer>=0.9",
]

[project.optional-dependencies]
llm = ["google-genai>=0.3"]
dev = ["pytest>=7", "pytest-mock>=3"]

[project.scripts]
citecheck = "citecheck.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/citecheck"]
